Solución de conflictos de types
