+++ proyecto/src/Types/usuariosTypes.js
@@ -1,8 +1,10 @@
-export const CAMBIAR_USUARIO_NOMBRE = 'cambiarUsuarioNombre';
-export const CAMBIAR_USUARIO_APELLIDO_PATERNO = 'cambiarUsuarioApellidoPaterno';
-export const CAMBIAR_USUARIO_APELLIDO_MATERNO = 'cambiarUsuarioApellidoMaterno';
-export const CAMBIAR_USUARIO_EDAD = 'cambiarUsuarioEdad';
-export const ERROR = 'error';
-export const INICIANDO = 'iniciando';
+export const CAMBIO_NOMBRE = 'cambiarUsuarioNombre';
+export const CAMBIO_APELLIDO_PATERNO = 'cambiarUsuarioApellidoPaterno';
+export const CAMBIO_APELLIDO_MATERNO = 'cambiarUsuarioApellidoMaterno';
+export const CAMBIO_EDAD = 'cambiarUsuarioEdad';
+export const ERROR_USUARIOS = 'error';
+export const EMPEZAR_USUARIOS = 'iniciando';
+export const CARGANDO_USUARIOS = 'iniciando';
 export const USUARIOS_CARGADOS = 'usuariosCargados';
 export const CONSULTA_USUARIOS = 'consultaUsuarios';
+export const USUARIO_SOLO_LECTURA = 'usuarioSoloLectura';
